otp_phone_auth/django_backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
firebase-admin==6.4.0
django-cors-headers==4.3.1
python-decouple==3.8
